+++ feature-staking-impl/src/main/java/jp/co/soramitsu/feature_staking_impl/presentation/validators/change/recommended/di/RecommendedValidatorsComponent.kt
@@ -1,10 +1,10 @@
-package jp.co.soramitsu.feature_staking_impl.presentation.validators.recommended.di
+package jp.co.soramitsu.feature_staking_impl.presentation.validators.change.recommended.di
 
 import androidx.fragment.app.Fragment
 import dagger.BindsInstance
 import dagger.Subcomponent
 import jp.co.soramitsu.common.di.scope.ScreenScope
-import jp.co.soramitsu.feature_staking_impl.presentation.validators.recommended.RecommendedValidatorsFragment
+import jp.co.soramitsu.feature_staking_impl.presentation.validators.change.recommended.RecommendedValidatorsFragment
 
 @Subcomponent(
     modules = [
